+++ diesel_compile_tests/tests/compile-fail/find_requires_correct_type.rs
@@ -28,5 +28,4 @@ fn main() {
     //~| ERROR E0277
     //~| ERROR E0277
     //~| ERROR E0277
-    //~| ERROR E0277
 }