+++ diesel_tests/tests/types.rs
@@ -491,7 +491,7 @@ use diesel::query_builder::{QueryFragment, QueryId};
 fn query_to_sql_equality<T, U>(sql_str: &str, value: U) -> bool where
     TestBackend: HasSqlType<T>,
     U: AsExpression<T> + Debug + Clone,
-    U::Expression: SelectableExpression<(), T> + QueryFragment<TestBackend> + QueryId,
+    U::Expression: SelectableExpression<()> + QueryFragment<TestBackend> + QueryId,
     T: QueryId,
 {
     use diesel::expression::dsl::sql;